Creating bookShelf

+++ book/src/main/java/br/com/bookon/book/repositories/BookRepository.java
@@ -1,7 +1,7 @@
 package br.com.bookon.book.repositories;
 
 
-import br.com.bookon.book.models.mongo.Book;
+import br.com.bookon.book.models.Book;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
@@ -13,9 +13,7 @@ public interface BookRepository extends MongoRepository<Book, String> {
 
     Book findBookById(String id);
 
-    List<Book> findBookByUserId(Long userId);
+    List<Book> findBookByUserId(Integer userId);
 
-    List<Book> findBooksByUserIdIn(List<Long> userIds);
-
-
+    List<Book> findBooksByUserIdIn(List<Integer> userIds);
 }
